+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+### 13.8.0
+
+- Allow arbitrary campaign keys
+
+### 13.7.1
+
+- Ignore `displayData` on `BapiClient.addOrUpdateItems` when strategy `ExistingItemHandling.AddQuantityToExisting` is used
+  - Reason: The underlying API does not support updating display data of existing items, only the custom data
+
+### 13.7.0
+
+- Support parameters `containsSearch` and `disableFuzziness` for product searches
+
 ### 13.6.0
 
 - `BapiClient.basket.updateItem` (and its usage through `BapiClient.basket.addOrUpdateItems`) now supports updating the `customData` as well as the `pricePromotionKey`